requirements_froze.txt
aenum==2.2.6
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fbchat==1.9.7
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-api-keys==0.5.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
paho-mqtt==1.6.1
pathspec==0.11.2
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.0.7
